add setup command for shell alias
The application cannot change the directory on the current shell
session, as it would be needed by the `go` subcommand.
To enable the feature, we implement the following
- the `go` subcommand emit the path of the directory to go to, prefixed
  with "goto::"
- a `wok` function is added in the shell; such function works as an
  wrapper for the real `wok` implementation
- the only purpose of the wrapper is to intercept "goto::" string and
  move directory accordingly

This commit prepare the scripts that implements the shell alias.
Furthermore, it add a `setup` subprogram to `wok` that installs such a
scripts in the current running directory.

(`go` subcommand to be implemented, yet)