add go command to move to project directory
The `go` subcommand move the current shell in the project's root
directory.

The search parameter is the organization/project string that identifies
the project.
A naive fuzzy search is implemented so that a partial string can be
provided. If there is only one match, the `go` command proceeds,
otherwise a list of possible matches is provided.

The `go` command per se does not change the shell directory (it cannot);
the target directory is printed to stdin, along with the well-known
GOTO_MARKER prefix.
The shell wrapper (previously installed via the `setup` subcommand)
takes care of intercept and parse the string and change the shell's
directory.