use file descriptor 3 to change directory
Refine the work of c2dd0007197fa65d8c2469ab6f777e0ab3ffebc7.
The "goto::" instruction is now passed on the file descriptor 3 instead
of on stdout. The shell wrapper catch it and change directory
accordingly.
At the same time, the full path of the directory of the found project is
printed to stdout.

Thanks to this modification, the shell wrapper does not have to retain
the `wok` output and can pipe it directly into stdout, allowing:
- for the `go` command, the destination directory is printed in stdout,
  allowing for composition with other programs
- command that have rich TUI elements using ASCII codes (i.e. spinners),
  they now render correctly