do not print destination folder
On the `go` command, destination folder should not be printed when
actually changing directory.
On the other hand, when piping the program to another, it is needed to
that the target program can read the directory path from its stdin.

Example:

`wok balanza/wok`: the directory is changed in the current shell,
nothing is printend.

`wok balanza/wok | xargs nvim $1`: open the project in NeoVim using the
full path as target